Add Has to Tx.
When working with very large data sets it is desirable to have a `Has`
function that does not require to touch disk for performance reasons.
Additionally, this also can keep memory usage down since value does not
need to be cached.

In order for this feature to be useful you must enable the
`HintKeyAndRAMIdxMode` option.